+++ lib/src/command/update_command.dart
@@ -1,7 +1,7 @@
 import 'package:args/command_runner.dart';
 import 'package:slidy/slidy.dart';
 
-class UpdateCommand extends Command {
+class UpdateCommand extends CommandBase {
   final name = "update";
   final description = "Update a new package or packages.";
 
@@ -12,7 +12,7 @@ class UpdateCommand extends Command {
 
   void run() {
     if (argResults.rest.isEmpty) {
-      throw new UsageException("value not passed for a module command", usage);
+      throw UsageException("value not passed for a module command", usage);
     } else {
       update(argResults.rest, argResults["dev"]);
     }