+++ lib/src/modules/install.dart
@@ -6,7 +6,7 @@ import 'package:slidy/src/services/pub_service.dart';
 import 'package:slidy/src/utils/utils.dart';
 import 'package:slidy/src/utils/output_utils.dart' as output;
 
-void install(List<String> packs, bool isDev) async {
+void install(List<String> packs, bool isDev, {bool haveTwoLines}) async {
   PubSpec spec = await getPubSpec();
   File yaml = File("pubspec.yaml");
   var node = yaml.readAsLinesSync();
@@ -14,6 +14,7 @@ void install(List<String> packs, bool isDev) async {
   int indexDependencyDev =
       node.indexWhere((t) => t.contains("dev_dependencies:")) + 1;
   bool isAlter = false;
+  haveTwoLines = haveTwoLines ?? false;
   var dependencies = isDev ? spec.devDependencies : spec.dependencies;
 
   for (String pack in packs) {
@@ -22,20 +23,27 @@ void install(List<String> packs, bool isDev) async {
 
     if (pack.contains(":")) {
       packName = pack.split(":")[0];
-      version = pack.split(":")[1];
+      version = pack.split(":").length > 1 
+        ? pack.split(":")[1] + ':' +  pack.split(":")[2] 
+        : pack.split(":")[1];
     } else {
       packName = pack;
     }
 
-    if (dependencies.containsKey(packName)) {
+    if (dependencies.containsKey(packName) && !haveTwoLines) {
       await update([packName], isDev);
       continue;
     }
 
     try {
-      version = await PubService().getPackage(packName, version);
-      node.insert(isDev ? indexDependencyDev : indexDependency,
-          "  $packName: ^$version");
+      if(!haveTwoLines){
+        version = await PubService().getPackage(packName, version);
+        node.insert(isDev ? indexDependencyDev : indexDependency,
+            "  $packName: ^$version");
+      } else if(!dependencies.containsKey(packName)){
+        node.insert(isDev ? indexDependencyDev : indexDependency,
+            "  $packName: \n    $version");
+      }
       output.success("$packName:$version Added in pubspec");
       isAlter = true;
     } catch (e) {